chore(parser): move reserved words to openapi
