+++ src/openApi/v2/parser/getModels.ts
@@ -1,5 +1,5 @@
 import type { Model } from '../../../types/client';
-import { reservedWords } from '../../../utils/reservedWords';
+import { reservedWords } from '../../common/parser/reservedWords';
 import { getType } from '../../common/parser/type';
 import type { OpenApi } from '../interfaces/OpenApi';
 import { getModel } from './getModel';
